fix(panel): avoid being stuck "loading" (#6)
Whenever the Copilot LSP knows it's not going to provide any
suggestions, it replies with PanelSolutionsDone even before it completes
the request. This resulted in us setting the status to "done" and then
immediately resetting it to "loading", and thus getting stuck on
"loading".

The fix is to set the status to "loading" in the initial refresh call
itself to avoid the race between on_solutions_done and the
get_panel_completions callback.

Co-authored-by: Tomas Janousek <tomi@nomi.cz>